fix(select): reset focus on input if dropdown is closed
