+++ packages/ui/src/composables/useFocusDeep.ts
@@ -1,31 +1,22 @@
-import { shallowRef, computed, Ref, onMounted, DefineComponent } from 'vue'
+import { computed, Ref, DefineComponent } from 'vue'
 import { useHTMLElement } from './useHTMLElement'
-import { useCaptureEvent } from './useCaptureEvent'
 import { useCurrentElement } from './useCurrentElement'
-import { blurElement, focusElement } from '../utils/focus'
+import { useFocusedElement } from './useFocusedElement'
 
-const useActiveElement = () => {
-  const activeEl = shallowRef<HTMLElement>()
-
-  const updateActiveElement = () => {
-    activeEl.value = document.activeElement as HTMLElement
-  }
-
-  onMounted(updateActiveElement)
-
-  useCaptureEvent('focus', updateActiveElement)
-  useCaptureEvent('blur', updateActiveElement)
-
-  return activeEl
-}
-
+/**
+ * `true` if `el` or any of his children are in focus
+ *
+ *  if set to `true` set `el` focused, but if any of `el` children was focused before, set child focused instead
+ *
+ * @notice this will not trigger native `focus` event and you need to trigger it manually and handle infinite loop
+ */
 export const useFocusDeep = (el?: Ref<HTMLElement | DefineComponent | undefined>) => {
-  const focused = useActiveElement()
+  const focused = useFocusedElement()
   const current = useCurrentElement(el ? useHTMLElement(el) : undefined)
   // Cache previouslyFocusedElement, so we can simply come back to it
   let previouslyFocusedElement: HTMLElement | null = null
 
-  return computed<boolean | undefined>({
+  const isFocused = computed<boolean | undefined>({
     get () {
       if (!focused.value) { return false }
       if (focused.value === current.value) { return true }
@@ -41,10 +32,30 @@ export const useFocusDeep = (el?: Ref<HTMLElement | DefineComponent | undefined>
         target = current.value
       }
 
+      // NOTICE: Focus and blur events will not be dispatched here to prevent infinite loop
       if (value) {
         (target)?.focus()
       } else {
         (target)?.blur()
+      }
+    },
+  })
+
+  return Object.assign(isFocused, {
+    /** Focus `el` if focus is not set to any other element */
+    focusIfNothingIfFocused: () => {
+      // body if focused by default, but we assume it means nothing is focused
+      // by nothing we mean elements like input, button, etc.
+      if (focused.value === document.body) {
+        isFocused.value = true
+      }
+    },
+
+    focusPreviousElement: () => {
+      if (previouslyFocusedElement) {
+        previouslyFocusedElement.focus()
+      } else {
+        document.body.focus()
       }
     },
   })
